requirements.txt
dominate==2.2.0
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
Flask-WTF==0.12
gunicorn==19.5.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.2.0
visitor==0.1.2
Werkzeug==0.11.9
WTForms==2.1
